+++ airsonic-main/src/main/java/org/airsonic/player/domain/PodcastChannel.java
@@ -23,6 +23,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import javax.persistence.*;
 
+import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -34,7 +35,7 @@ import java.util.List;
  */
 @Entity
 @Table(name = "podcast_channel")
-public class PodcastChannel {
+public class PodcastChannel implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -151,4 +152,5 @@ public class PodcastChannel {
     public MediaFile getMediaFile() {
         return mediaFile;
     }
+
 }
