+++ airsonic-main/src/main/java/org/airsonic/player/repository/PodcastChannelRepository.java
@@ -1,8 +1,15 @@
 package org.airsonic.player.repository;
 
 import org.airsonic.player.domain.PodcastChannel;
+import org.airsonic.player.domain.PodcastStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PodcastChannelRepository extends JpaRepository<PodcastChannel, Integer> {}
+public interface PodcastChannelRepository extends JpaRepository<PodcastChannel, Integer> {
+
+    List<PodcastChannel> findByStatus(PodcastStatus status);
+
+}
