+++ airsonic-main/src/test/resources/application.properties
@@ -16,3 +16,4 @@ spring.datasource.username=sa
 spring.datasource.password=
 spring.liquibase.change-log=classpath:liquibase/db-changelog.xml
 spring.jpa.open-in-view=false
+spring.main.allow-bean-definition-overriding=true
